1) Added first dirty implementation to pretty print the confusion matrix 2) Added converter to cast fields in a matrix, that makes illogical casts in kNN when calculating the difference between two values obsolete 3) Added tree learning method, decision stump, with support for more. Tested on sample dataset 4) Added static factory methods to the matrix structures to generate a matrix of certain size, however since the indexing structure in combination with setting a value is far from ideal at the moment it is almost always necessary to call get_raw() on the newly generated matrix right away and work on the raw form when filling the matrix.
Known limitation:
At the moment the indexing is comparable to the normal indexing system of multi-dimensional arrays. However, the line `matrix[i][j] = value` is executed by first resolving matrix[i], by calling __getitem__, and then calling __setitem__ on the result. At the moment __getitem__ of a Matrix2D or Vector results in a Matrix2D or Vector object created with the raw data that was requested. That means that the __setitem__ is called on a new object that has nothing to do with matrix[i] anymore, therefore resulting in the situation that any modifications are done on the raw data within the new object and not in matrix.

That means that when you run the following example, matrix won't be filled with nonzero elements:
>matrix = Matrix2D.genmat(10,10,0)
>for i in range(10):
>  for j in range(10):
>    matrix[i][j] = i+j

The same goes for the following example:
>matrix = Matrix2D.genmat(10,10,0)
>for row in matrix:
>  for c in range(10):
>    row[c] = c

The problem seems to be inherent with the interpretation of object[index1][index2], so perhaps a "good" solution would be to work with a different indexing system. A possible alternative is to use matrix[dim1, dim2, dim3, dim....] to point to matrix[dim1][dim2][dim3][dim...]. That means that the current interpretation of this input, which is a list of indexes on the same dimension, should change. This could possibly become:

matrix[i]       = all other dimensions for dim1 = i
matrix[i,j]     = all other dimensions for dim1 = i and dim2 = j
matrix[(i,j)]   = equivalent to the above
matrix[[i,j]]   = all other dimensions for dim1 = i or j
matrix[i:j]     = all other dimensions for dim1 = i....j
matrix[i:j,m:n] = all other dimensions for dim1 = i....j and dim2 = m....n (if this is even possible)